refactor(frontend/library): extract briefing tab sub-components and reorganize panel
Split BriefingTabContent into focused components under a local components/
folder, move AgentBriefingPanel children into components/, extract SitrepItem
helpers, and remove unused SitrepList.
