--- autogpt_platform/frontend/src/app/(platform)/library/components/AgentBriefingPanel/AgentBriefingPanel.module.css
@@ -1,44 +1,0 @@
-.glassPanel {
-  position: relative;
-  isolation: isolate;
-}
-
-.glassPanel::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  border-radius: inherit;
-  padding: 1px;
-  background: conic-gradient(
-    from var(--border-angle, 0deg),
-    rgba(129, 120, 228, 0.04),
-    rgba(129, 120, 228, 0.14),
-    rgba(168, 130, 255, 0.09),
-    rgba(129, 120, 228, 0.04),
-    rgba(99, 102, 241, 0.12),
-    rgba(129, 120, 228, 0.04)
-  );
-  -webkit-mask:
-    linear-gradient(#000 0 0) content-box,
-    linear-gradient(#000 0 0);
-  mask:
-    linear-gradient(#000 0 0) content-box,
-    linear-gradient(#000 0 0);
-  -webkit-mask-composite: xor;
-  mask-composite: exclude;
-  animation: rotate-border 6s linear infinite;
-  pointer-events: none;
-  z-index: -1;
-}
-
-@property --border-angle {
-  syntax: "<angle>";
-  initial-value: 0deg;
-  inherits: false;
-}
-
-@keyframes rotate-border {
-  to {
-    --border-angle: 360deg;
-  }
-}
