+++ autogpt_platform/frontend/src/app/(platform)/library/components/AgentBriefingPanel/components/BriefingTabContent/__tests__/BriefingTabContent.test.tsx
@@ -196,7 +196,7 @@ describe("BriefingTabContent — CostsBreakdown", () => {
 
     // Headline stats
     expect(screen.getByText("$42.50")).toBeDefined();
-    expect(screen.getByText("Most expensive runs")).toBeDefined();
+    expect(screen.getByText("Most expensive tasks")).toBeDefined();
     expect(screen.getByText("Spend by agent")).toBeDefined();
 
     // Agent names resolved via graph_id lookup
