+++ autogpt_platform/frontend/src/app/(platform)/library/components/AgentBriefingPanel/components/CostsBreakdown/components/SpendByAgentList.tsx
