+++ autogpt_platform/frontend/src/app/(platform)/library/components/AgentBriefingPanel/components/CostsBreakdown/components/TopRunsList.tsx
@@ -25,7 +25,7 @@ export function TopRunsList({ runs, agentLookup }: Props) {
   return (
     <section className="flex flex-col gap-2 lg:mt-[1rem]">
       <Text variant="body-medium" className="text-neutral-800 lg:mb-[.5rem]">
-        Most expensive runs
+        Most expensive tasks
       </Text>
       <ul className="flex flex-col divide-y divide-zinc-100 rounded-medium border border-zinc-100 bg-white">
         {visible.map((run) => {
