+++ autogpt_platform/frontend/src/app/(platform)/library/components/SitrepItem/SitrepItem.tsx
@@ -18,6 +18,7 @@ import { useRouter } from "next/navigation";
 import type { SitrepItemData, SitrepPriority } from "../../types";
 import { ContextualActionButton } from "../ContextualActionButton/ContextualActionButton";
 import styles from "./SitrepItem.module.css";
+import { buildAutoPilotPrompt } from "./helpers";
 
 interface Props {
   item: SitrepItemData;
@@ -153,23 +154,4 @@ export function SitrepItem({ item }: Props) {
       </div>
     </div>
   );
-}
-
-function buildAutoPilotPrompt(item: SitrepItemData): string {
-  switch (item.priority) {
-    case "error":
-      return `What happened with ${item.agentName}? It says "${item.message}" — can you check the logs and tell me what to fix?`;
-    case "running":
-      return `Give me a status update on the ${item.agentName} run — what has it found so far?`;
-    case "stale":
-      return `${item.agentName} hasn't run recently. Should I keep it or update and re-run it?`;
-    case "success":
-      return `Show me what ${item.agentName} found in its last run — summarize the results and any key takeaways.`;
-    case "listening":
-      return `What is ${item.agentName} listening for? Give me a summary of its trigger configuration.`;
-    case "scheduled":
-      return `When is ${item.agentName} scheduled to run next?`;
-    case "idle":
-      return `${item.agentName} has been idle. Should I keep it or update and re-run it?`;
-  }
 }
